backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28",
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
